% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alkahest-package.R
\docType{package}
\name{alkahest-package}
\alias{alkahest-package}
\alias{_PACKAGE}
\alias{alkahest}
\title{alkahest: Pre-Processing XY Data from Experimental Methods}
\description{
A lightweight, dependency-free toolbox for pre-processing XY data from experimental methods (i.e. any signal that can be measured along a continuous variable). This package provides methods for baseline estimation and correction, smoothing, normalization, integration and peaks detection.
}
\details{
\tabular{ll}{
\strong{Package:} \tab alkahest \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab  \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/alkahest/}
  \item \url{https://github.com/tesselle/alkahest}
  \item Report bugs at \url{https://github.com/tesselle/alkahest/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
