% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/baseline.R
\docType{methods}
\name{baseline}
\alias{baseline}
\alias{baseline_linear}
\alias{baseline_linear-method}
\alias{baseline_rubberband}
\alias{baseline_rubberband-method}
\alias{baseline_snip}
\alias{baseline_snip-method}
\alias{baseline_linear,numeric,numeric-method}
\alias{baseline_linear,ANY,missing-method}
\alias{baseline_rubberband,numeric,numeric-method}
\alias{baseline_rubberband,ANY,missing-method}
\alias{baseline_snip,numeric,numeric-method}
\alias{baseline_snip,ANY,missing-method}
\title{Baseline Estimation}
\usage{
baseline_linear(x, y, ...)

baseline_rubberband(x, y, ...)

baseline_snip(x, y, ...)

\S4method{baseline_linear}{numeric,numeric}(x, y, from = min(x), to = max(x))

\S4method{baseline_linear}{ANY,missing}(x, from = min(x), to = max(x))

\S4method{baseline_rubberband}{numeric,numeric}(x, y, noise = 0, spline = TRUE, ...)

\S4method{baseline_rubberband}{ANY,missing}(x, noise = 0, spline = TRUE, ...)

\S4method{baseline_snip}{numeric,numeric}(x, y, LLS = FALSE, decreasing = FALSE, n = 100, ...)

\S4method{baseline_snip}{ANY,missing}(x, LLS = FALSE, decreasing = FALSE, n = 100, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra parameters to be passed to further methods.}

\item{from}{An \code{\link{numeric}} value giving the first data point (in \code{x} unit)
to be used for linear interpolation.}

\item{to}{An \code{\link{integer}} value giving the last data point (in \code{x} unit)
to be used for linear interpolation.}

\item{noise}{A length-one \code{\link{numeric}} vector giving the noise level.
Only used if \code{method} is "\code{rubberband}".}

\item{spline}{A \code{\link{logical}} scalar: should spline interpolation through the
support points be used instead of linear interpolation?
Only used if \code{method} is "\code{rubberband}".}

\item{LLS}{A \code{\link{logical}} scalar: should the LLS operator be applied on \code{x}
before employing SNIP algorithm? Only used if \code{method} is "\code{SNIP}".}

\item{decreasing}{A \code{\link{logical}} scalar: should a decreasing clipping window
be used? Only used if \code{method} is "\code{SNIP}".}

\item{n}{An \code{\link{integer}} value giving the number of iterations.
Only used if \code{method} is "\code{SNIP}".}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Baseline Estimation
}
\details{
The following methods are available for baseline estimation:
\describe{
\item{\code{SNIP}}{Sensitive Nonlinear Iterative Peak clipping algorithm.}
\item{\code{rubberband}}{A convex envelope of the spectrum is determined and the
baseline is estimated as the part of the convex envelope lying below the
spectrum. Note that the rubber band does not enter the concave regions
(if any) of the spectrum.}
\item{\code{linear}}{Linear baseline estimation.}
}
}
\note{
\code{baseline_rubberband()} is slightly modified from C. Beleites'
\code{\link[hyperSpec:spc-rubberband]{hyperSpec::spc.rubberband()}}.
}
\examples{
## gamma-ray spectrometry
data("BEGe")

## Subset from 2.75 to 200 keV
BEGe <- signal_select(BEGe, from = 3, to = 200)

## Plot spectrum
plot(BEGe, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Linear baseline
BEGe_linear <- baseline_linear(x = BEGe, from = 8, to = 20)

plot(BEGe, type = "l", xlab = "Energy (keV)", ylab = "Count")
lines(BEGe_linear, type = "l", col = "red")

## Rubberband baseline
BEGe_rubber <- baseline_rubberband(BEGe)

plot(BEGe, type = "l", xlab = "Energy (keV)", ylab = "Count")
lines(BEGe_rubber, type = "l", col = "red")

## SNIP baseline
BEGe_snip <- baseline_snip(BEGe, LLS = FALSE, decreasing = FALSE, n = 100)

plot(BEGe, type = "l", xlab = "Energy (keV)", ylab = "Count")
lines(BEGe_snip, type = "l", col = "red")
}
\references{
Liland, K. H. (2015). 4S Peak Filling - baseline estimation by iterative
mean suppression. \emph{MethodsX}, 2, 135-140. \doi{10.1016/j.mex.2015.02.009}.

Morháč, M., Kliman, J., Matoušek, V., Veselský, M. & Turzo, I. (1997).
Background elimination methods for multidimensional gamma-ray spectra.
\emph{Nuclear Instruments and Methods in Physics Research Section A:
Accelerators, Spectrometers, Detectors and Associated Equipment}, 401(1),
p. 113-132. \doi{10.1016/S0168-9002(97)01023-1}

Morháč, M. & Matoušek, V. (2008). Peak Clipping Algorithms for Background
Estimation in Spectroscopic Data. \emph{Applied Spectroscopy}, 62(1), p. 91-106.
\doi{10.1366/000370208783412762}

Ryan, C. G., Clayton, E., Griffin, W. L., Sie, S. H. & Cousens, D. R.
(1988). SNIP, a statistics-sensitive background treatment for the
quantitative analysis of PIXE spectra in geoscience applications.
\emph{Nuclear Instruments and Methods in Physics Research Section B:
Beam Interactions with Materials and Atoms}, 34(3), p. 396-402.
\doi{10.1016/0168-583X(88)90063-8}
}
\author{
N. Frerebeau
}
\concept{baseline estimation methods}
