% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BEGe}
\alias{BEGe}
\title{Gamma-Ray Spectrometry}
\format{
A \code{\link{data.frame}} with 8192 rows (channels) and 2 variables.
\describe{
\item{energy}{(keV)}
\item{count}{}
}
}
\usage{
BEGe
}
\description{
Gamma-Ray Spectrometry
}
\examples{
data("BEGe")
plot(BEGe, type = "l", xlab = "Energy (keV)", ylab = "Count")
}
\seealso{
Other datasets: 
\code{\link{LaBr}},
\code{\link{Raman}},
\code{\link{XRD}}
}
\concept{datasets}
\keyword{datasets}
