% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/integrate.R
\docType{methods}
\name{integrate_trapezoid}
\alias{integrate_trapezoid}
\alias{signal_integrate-method}
\alias{integrate_trapezoid,numeric,numeric-method}
\title{Trapezoidal Rule}
\usage{
integrate_trapezoid(x, y, ...)

\S4method{integrate_trapezoid}{numeric,numeric}(x, y)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Approximates the definite integral by using the trapezoidal rule.
}
\examples{
## Calculate the area under the sine curve from 0 to pi
x <- seq(0, pi, len = 101)
y <- sin(x)
integrate_trapezoid(x, y) # 1.999835504
}
\author{
N. Frerebeau
}
\concept{integration methods}
