% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirke-package.R
\docType{package}
\name{kirke-package}
\alias{kirke-package}
\alias{_PACKAGE}
\alias{kirke}
\title{kirke: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
\tabular{ll}{
\strong{Package:} \tab kirke \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 0.0.0.9000 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab  \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/kirke/}
  \item \url{https://github.com/tesselle/kirke}
  \item Report bugs at \url{https://github.com/tesselle/kirke/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
