% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaBr}
\alias{LaBr}
\title{In-situ Gamma-Ray Spectrometry}
\format{
A \code{\link{data.frame}} with 1024 rows (channels) and 2 variables.
\describe{
\item{energy}{(keV).}
\item{count}{}
}
}
\usage{
LaBr
}
\description{
In-situ Gamma-Ray Spectrometry
}
\seealso{
Other datasets: 
\code{\link{BEGe}}
}
\concept{datasets}
\keyword{datasets}
