% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/peaks.R
\docType{methods}
\name{peaks_find}
\alias{peaks_find}
\alias{peaks_find-method}
\alias{peaks_find,numeric,numeric-method}
\alias{peaks_find,ANY,missing-method}
\title{Find Peaks}
\usage{
peaks_find(x, y, ...)

\S4method{peaks_find}{numeric,numeric}(x, y, method = c("MAD"), SNR = 2, m = NULL, ...)

\S4method{peaks_find}{ANY,missing}(x, method = c("MAD"), SNR = 2, m = NULL, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra parameters to be passed to internal methods.}

\item{method}{A \code{\link{character}} string specifying the method to be used for
background noise estimation (see below).}

\item{SNR}{An \code{\link{integer}} giving the signal-to-noise-ratio for peak detection
(see below).}

\item{m}{An odd \code{\link{integer}} giving the half window size.
If \code{NULL}, 5\% of the data points is used as the half window size.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Finds local maxima in sequential data.
}
\details{
A local maximum has to be the highest one in the given window and has to be
higher than \eqn{SNR \times noise}{SNR * noise} to be recognized as peak.

The following methods are available for noise estimation:
\describe{
\item{\code{MAD}}{Median Absolute Deviation.}
}
}
\note{
Adapted from Stasia Grinberg's
\href{https://github.com/stas-g/findPeaks}{\code{findPeaks}} function.
}
\examples{
## gamma-ray spectrometry
data("BEGe")

## Subset from 2.75 to 200 keV
BEGe <- signal_select(BEGe, from = 3, to = 200)

## Correct baseline
BEGe_correct <- signal_correct(BEGe, method = "SNIP")

## Plot spectrum
plot(BEGe_correct, type = "l", xlab = "Energy (keV)", ylab = "Count")
abline(h = mad(BEGe_correct$y) * 3) # noise threshold

## Find peaks
BEGe_peaks <- peaks_find(BEGe_correct, SNR = 3, m = 5)
lines(BEGe_peaks, type = "p", pch = 16, col = "red")

## Half-Width at Half-Maximum
set.seed(12345)
x <- seq(-4, 4, length = 1000)
y <- dnorm(x)

peaks_fwhm(x, y, center = 0) # Expected: 2 * sqrt(2 * log(2))
}
\seealso{
Other peaks detection methods: 
\code{\link{peaks_fwhm}()}
}
\author{
N. Frerebeau
}
\concept{peaks detection methods}
