% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/peaks.R
\docType{methods}
\name{peaks_fwhm}
\alias{peaks_fwhm}
\alias{peaks_fwhm-method}
\alias{peaks_fwhm,numeric,numeric-method}
\alias{peaks_fwhm,ANY,missing-method}
\title{Half-Width at Half-Maximum}
\usage{
peaks_fwhm(x, y, ...)

\S4method{peaks_fwhm}{numeric,numeric}(x, y, center)

\S4method{peaks_fwhm}{ANY,missing}(x, center)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}

\item{center}{A \code{\link{numeric}} value giving the peak position in \code{x} units.}
}
\value{
A \code{\link{numeric}} value.
}
\description{
Estimates the Half-Width at Half-Maximum (FWHM) for a given peak.
}
\details{
It tries to get the smallest possible estimate.
}
\examples{
## gamma-ray spectrometry
data("BEGe")

## Subset from 2.75 to 200 keV
BEGe <- signal_select(BEGe, from = 3, to = 200)

## Correct baseline
BEGe_correct <- signal_correct(BEGe, method = "SNIP")

## Plot spectrum
plot(BEGe_correct, type = "l", xlab = "Energy (keV)", ylab = "Count")
abline(h = mad(BEGe_correct$y) * 3) # noise threshold

## Find peaks
BEGe_peaks <- peaks_find(BEGe_correct, SNR = 3, m = 5)
lines(BEGe_peaks, type = "p", pch = 16, col = "red")

## gamma-ray spectrometry
data("LaBr")

## Plot spectrum
LaBr <- signal_select(LaBr, from = 1350, to = 1650)
plot(LaBr, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Half-Width at Half-Maximum
peaks_fwhm(LaBr, center = 1490)
}
\seealso{
Other peaks detection methods: 
\code{\link{peaks_find}()}
}
\author{
N. Frerebeau
}
\concept{peaks detection methods}
