% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Raman}
\alias{Raman}
\title{Raman Spectroscopy}
\format{
A \code{\link{data.frame}} with 1182 rows and 2 variables.
\describe{
\item{shift}{Raman shift.}
\item{intensity}{}
}
}
\usage{
Raman
}
\description{
Raman Spectroscopy
}
\examples{
data("Raman")
plot(Raman, type = "l", xlab = "Shift", ylab = "Intensity")
}
\seealso{
Other datasets: 
\code{\link{BEGe}},
\code{\link{LaBr}},
\code{\link{XRD}}
}
\concept{datasets}
\keyword{datasets}
