% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/replace.R
\docType{methods}
\name{replace_negative}
\alias{replace_negative}
\alias{replace_negative-method}
\alias{replace_negative,numeric,numeric-method}
\alias{replace_negative,ANY,missing-method}
\title{Replace Negative Values}
\usage{
replace_negative(x, y, ...)

\S4method{replace_negative}{numeric,numeric}(x, y, value = 0)

\S4method{replace_negative}{ANY,missing}(x, value = 0)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra parameters to be passed to \code{threshold}.}

\item{value}{A \code{\link{numeric}} value to replace negative values.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Replace Negative Values
}
\seealso{
Other replacement methods: 
\code{\link{replace_threshold}()}
}
\author{
N. Frerebeau
}
\concept{replacement methods}
