% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/resample.R
\docType{methods}
\name{resample_bin}
\alias{resample_bin}
\alias{resample_bin-method}
\alias{resample_bin,numeric,numeric-method}
\alias{resample_bin,ANY,missing-method}
\title{Bin}
\usage{
resample_bin(x, y, ...)

\S4method{resample_bin}{numeric,numeric}(x, y, by)

\S4method{resample_bin}{ANY,missing}(x, y, by)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}

\item{by}{An \code{\link{integer}} specifying the binning ratio (i.e. the number of
points to be grouped together).}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Averages \code{x} values and sums the corresponding \code{y} values.
}
\examples{
## X-ray diffraction
data("XRD")

## Plot spectrum
plot(XRD, type = "l", xlab = expression(2*theta), ylab = "Count")

## Bin by 3
XRD_bin <- resample_bin(XRD, by = 3)
plot(XRD, type = "l", xlab = expression(2*theta), ylab = "Count")
lines(XRD_bin, type = "l", col = "red")

## Downsample by 10
XRD_down <- resample_down(XRD, by = 10)
plot(XRD, type = "l", xlim = c(20, 40), xlab = expression(2*theta), ylab = "Count")
lines(XRD_down, type = "l", col = "red")

## Linearly interpolate
XRD_approx <- resample_interpolate(XRD, from = 20, to = 40, by = 0.02)
plot(XRD, type = "l", xlim = c(20, 40), xlab = expression(2*theta), ylab = "Count")
lines(XRD_approx, type = "l", col = "red")
}
\seealso{
Other resampling methods: 
\code{\link{resample_down}()},
\code{\link{resample_interpolate}()}
}
\author{
N. Frerebeau
}
\concept{resampling methods}
