% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/shift.R
\docType{methods}
\name{shift_interpolate}
\alias{shift_interpolate}
\alias{shift_interpolate-method}
\alias{shift_interpolate,numeric,numeric-method}
\alias{shift_interpolate,ANY,missing-method}
\title{Linearly Interpolate a New X Scale}
\usage{
shift_interpolate(x, y, ...)

\S4method{shift_interpolate}{numeric,numeric}(x, y, from, to, by, ...)

\S4method{shift_interpolate}{ANY,missing}(x, y, from, to, by, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra arguments to be passed to \code{\link[stats:approxfun]{stats::approx()}}.}

\item{from}{A length-one \code{\link{numeric}} vector giving the starting value of the
sequence where interpolation is to take place.}

\item{to}{A length-one \code{\link{numeric}} vector giving the end value of the
sequence where interpolation is to take place.}

\item{by}{A length-one \code{\link{numeric}} vector specifying the increment of the
sequence.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Linearly Interpolate a New X Scale
}
\examples{
## X-ray diffraction
data("XRD")

## Plot spectrum
plot(XRD, type = "l", xlab = expression(2*theta), ylab = "Count")

## Shift by one degree
XRD_offset <- shift_offset(XRD, delta = 1)
lines(XRD_offset, type = "l", col = "red")

## Linearly interpolate a new x scale
XRD_approx <- shift_interpolate(XRD, from = 20, to = 40, by = 0.02)
plot(XRD, type = "l", xlim = c(20, 40), xlab = expression(2*theta), ylab = "Count")
lines(XRD_approx, type = "l", col = "red")
}
\seealso{
Other shifting methods: 
\code{\link{shift_offset}()}
}
\author{
N. Frerebeau
}
\concept{shifting methods}
