% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/shift.R
\docType{methods}
\name{shift_offset}
\alias{shift_offset}
\alias{shift_offset-method}
\alias{shift_offset,numeric,numeric-method}
\alias{shift_offset,ANY,missing-method}
\title{Shift the X Scale by a Given Value}
\usage{
shift_offset(x, y, ...)

\S4method{shift_offset}{numeric,numeric}(x, y, delta)

\S4method{shift_offset}{ANY,missing}(x, y, delta)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}

\item{delta}{A length-one \code{\link{numeric}} vector specifying the offset.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Shift the X Scale by a Given Value
}
\examples{
## X-ray diffraction
data("XRD")

## Plot spectrum
plot(XRD, type = "l", xlab = expression(2*theta), ylab = "Count")

## Shift by one degree
XRD_offset <- shift_offset(XRD, delta = 1)
lines(XRD_offset, type = "l", col = "red")

## Linearly interpolate a new x scale
XRD_approx <- shift_interpolate(XRD, from = 20, to = 40, by = 0.02)
plot(XRD, type = "l", xlim = c(20, 40), xlab = expression(2*theta), ylab = "Count")
lines(XRD_approx, type = "l", col = "red")
}
\seealso{
Other shifting methods: 
\code{\link{shift_interpolate}()}
}
\author{
N. Frerebeau
}
\concept{shifting methods}
