% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/signal.R
\docType{methods}
\name{signal_correct}
\alias{signal_correct}
\alias{signal_correct-method}
\alias{signal_correct,numeric,numeric-method}
\alias{signal_correct,ANY,missing-method}
\title{Baseline Correction}
\usage{
signal_correct(x, y, ...)

\S4method{signal_correct}{numeric,numeric}(x, y, method = c("linear", "rubberband", "SNIP"), ...)

\S4method{signal_correct}{ANY,missing}(x, method = c("linear", "rubberband", "SNIP"), ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra arguments to be passed to \verb{baseline_*()}.}

\item{method}{A \code{\link{character}} string specifying the method for baseline
estimation. It must be one of "\code{linear}", "\code{rubberband}" or "\code{SNIP}".
Any unambiguous substring can be given.}
}
\description{
Baseline Correction
}
\examples{
## gamma-ray spectrometry
data("BEGe")

## Subset from 2.75 to 200 keV
BEGe <- signal_select(BEGe, from = 3, to = 200)

## Baseline correction
BEGe_corr <- signal_correct(BEGe, method = "SNIP")

plot(BEGe, type = "l", xlab = "Energy (keV)", ylab = "Count")
lines(BEGe_corr, type = "l", col = "red")
}
\seealso{
Other signal processing methods: 
\code{\link{signal_shift}()},
\code{\link{subset}()}
}
\concept{signal processing methods}
