% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/signal.R
\docType{methods}
\name{signal_mean}
\alias{signal_mean}
\alias{signal_mean-method}
\alias{signal_mean,ANY-method}
\title{Mean}
\usage{
signal_mean(...)

\S4method{signal_mean}{ANY}(...)
}
\arguments{
\item{...}{Any object that can be interpreted in a suitable way
(see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}
}
\description{
Mean
}
\examples{
## X-ray diffraction
data("XRD")

XRD1 <- XRD
XRD1$count <- XRD1$count + 1500

## Mean
XRD_mean <- signal_mean(XRD, XRD1)

plot(NULL, type = "l", xlim = c(10, 70) , ylim = c(3000, 36000),
     xlab = expression(2*theta), ylab = "Count")
lines(XRD, type = "l")
lines(XRD1, type = "l")
lines(XRD_mean, type = "l", col = "red")
}
\seealso{
Other signal processing methods: 
\code{\link{signal_correct}()},
\code{\link{signal_drift}()},
\code{\link{signal_shift}()},
\code{\link{subset}()}
}
\concept{signal processing methods}
