% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smooth.R
\docType{methods}
\name{smooth}
\alias{smooth}
\alias{smooth_rectangular}
\alias{smooth_rectangular-method}
\alias{smooth_triangular}
\alias{smooth_triangular-method}
\alias{smooth_savitzky}
\alias{smooth_savitzky-method}
\alias{smooth_rectangular,numeric,numeric-method}
\alias{smooth_rectangular,ANY,missing-method}
\alias{smooth_triangular,numeric,numeric-method}
\alias{smooth_triangular,ANY,missing-method}
\alias{smooth_savitzky,numeric,numeric-method}
\alias{smooth_savitzky,ANY,missing-method}
\title{Smooth}
\usage{
smooth_rectangular(x, y, ...)

smooth_triangular(x, y, ...)

smooth_savitzky(x, y, ...)

\S4method{smooth_rectangular}{numeric,numeric}(x, y, m = 3)

\S4method{smooth_rectangular}{ANY,missing}(x, m)

\S4method{smooth_triangular}{numeric,numeric}(x, y, m = 3)

\S4method{smooth_triangular}{ANY,missing}(x, m)

\S4method{smooth_savitzky}{numeric,numeric}(x, y, m = 3, p = 2)

\S4method{smooth_savitzky}{ANY,missing}(x, m, p)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}

\item{m}{An odd \code{\link{integer}} giving the number of adjacent points to be used.}

\item{p}{An \code{\link{integer}} giving the polynomial degree.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Smoothes intensities.
}
\details{
The following smoothing methods are available:
\describe{
\item{\code{rectangular}}{Unweighted sliding-average or rectangular smooth.
It replaces each point in the signal with the average of \eqn{m} adjacent
points.}
\item{\code{triangular}}{Weighted sliding-average or triangular smooth.
It replaces each point in the signal with the weighted mean of \eqn{m}
adjacent points.}
\item{\code{savitzky}}{Savitzky-Golay filter. This method is based on the
least-squares fitting of polynomials to segments of \eqn{m} adjacent
points.}
}
There will be \eqn{(m - 1) / 2} points both at the beginning and at the end
of the spectrum for which a complete \eqn{m}-width smooth cannot be
calculated. To prevent data loss, progressively smaller smooths are used at
the ends of the spectrum by \code{smooth_rectangular()} and
\code{smooth_triangular()}. If the Savitzky-Golay filter is used, the original
\eqn{(m - 1) / 2} points at the ends of the spectrum are preserved.
}
\examples{
## gamma-ray spectrometry
data("LaBr")

## Plot spectrum
LaBr <- signal_select(LaBr, from = 1350, to = 1650)
plot(LaBr, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Rectangular smooth
LaBr_unweighted <- smooth_rectangular(LaBr, m = 3)
plot(LaBr_unweighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Triangular smooth
LaBr_weighted <- smooth_triangular(LaBr, m = 5)
plot(LaBr_weighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Savitzky–Golay
LaBr_savitzky <- smooth_savitzky(LaBr, m = 21, p = 2)
plot(LaBr_savitzky, type = "l", xlab = "Energy (keV)", ylab = "Count")
}
\references{
Gorry, P. A. (1990). General Least-Squares Smoothing and Differentiation by
the Convolution (Savitzky-Golay) Method. \emph{Analytical Chemistry}, 62(6),
p. 570-573. \doi{10.1021/ac00205a007}.

Savitzky, A. & Golay, M. J. E. (1964). Smoothing and Differentiation of
Data by Simplified Least Squares Procedures. \emph{Analytical Chemistry},
36(8), p. 1627-1639. \doi{10.1021/ac60214a047}.
}
\seealso{
Other signal processing methods: 
\code{\link{signal_transform}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{signal processing methods}
