% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smooth.R
\docType{methods}
\name{smooth_loess}
\alias{smooth_loess}
\alias{smooth_loess-method}
\alias{smooth_loess,numeric,numeric-method}
\alias{smooth_loess,ANY,missing-method}
\title{Loess Smoothing}
\usage{
smooth_loess(x, y, ...)

\S4method{smooth_loess}{numeric,numeric}(x, y, span = 0.75, ...)

\S4method{smooth_loess}{ANY,missing}(x, span = 0.75, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra arguments to be passed to \code{\link[stats:loess]{stats::loess()}}.}

\item{span}{An \code{\link{integer}} specifying the degree of smoothing (see
\code{\link[stats:loess]{stats::loess()}}).}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Smoothes intensities by loess fitting.
}
\examples{
## gamma-ray spectrometry
data("LaBr")

## Plot spectrum
LaBr <- signal_select(LaBr, from = 1350, to = 1650)
plot(LaBr, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Rectangular smoothing
LaBr_unweighted <- smooth_rectangular(LaBr, m = 3)
plot(LaBr_unweighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Triangular smoothing
LaBr_weighted <- smooth_triangular(LaBr, m = 5)
plot(LaBr_weighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Loess smoothing
LaBr_loess <- smooth_loess(LaBr, span = 0.75)
plot(LaBr_loess, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Savitzky–Golay filter
LaBr_savitzky <- smooth_savitzky(LaBr, m = 21, p = 2)
plot(LaBr_savitzky, type = "l", xlab = "Energy (keV)", ylab = "Count")
}
\seealso{
Other smoothing methods: 
\code{\link{smooth_rectangular}()},
\code{\link{smooth_savitzky}()},
\code{\link{smooth_triangular}()}
}
\author{
N. Frerebeau
}
\concept{smoothing methods}
