% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smooth.R
\docType{methods}
\name{smooth_rectangular}
\alias{smooth_rectangular}
\alias{smooth_rectangular-method}
\alias{smooth_rectangular,numeric,numeric-method}
\alias{smooth_rectangular,ANY,missing-method}
\title{Rectangular Smoothing}
\usage{
smooth_rectangular(x, y, ...)

\S4method{smooth_rectangular}{numeric,numeric}(x, y, m = 3)

\S4method{smooth_rectangular}{ANY,missing}(x, m)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}

\item{m}{An odd \code{\link{integer}} giving the number of adjacent points to be used.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Unweighted sliding-average or rectangular Smoothing.
}
\details{
It replaces each point in the signal with the average of \eqn{m} adjacent
points.

There will be \eqn{(m - 1) / 2} points both at the beginning and at the end
of the spectrum for which a complete \eqn{m}-width smooth cannot be
calculated. To prevent data loss, progressively smaller smooths are used at
the ends of the spectrum.
}
\examples{
## gamma-ray spectrometry
data("LaBr")

## Plot spectrum
LaBr <- signal_select(LaBr, from = 1350, to = 1650)
plot(LaBr, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Rectangular smoothing
LaBr_unweighted <- smooth_rectangular(LaBr, m = 3)
plot(LaBr_unweighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Triangular smoothing
LaBr_weighted <- smooth_triangular(LaBr, m = 5)
plot(LaBr_weighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Loess smoothing
LaBr_loess <- smooth_loess(LaBr, span = 0.75)
plot(LaBr_loess, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Savitzky–Golay filter
LaBr_savitzky <- smooth_savitzky(LaBr, m = 21, p = 2)
plot(LaBr_savitzky, type = "l", xlab = "Energy (keV)", ylab = "Count")
}
\seealso{
Other smoothing methods: 
\code{\link{smooth_loess}()},
\code{\link{smooth_savitzky}()},
\code{\link{smooth_triangular}()}
}
\author{
N. Frerebeau
}
\concept{smoothing methods}
