% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/smooth.R
\docType{methods}
\name{smooth_savitzky}
\alias{smooth_savitzky}
\alias{smooth_savitzky-method}
\alias{smooth_savitzky,numeric,numeric-method}
\alias{smooth_savitzky,ANY,missing-method}
\title{Savitzky-Golay Filter}
\usage{
smooth_savitzky(x, y, ...)

\S4method{smooth_savitzky}{numeric,numeric}(x, y, m = 3, p = 2)

\S4method{smooth_savitzky}{ANY,missing}(x, m, p)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Currently not used.}

\item{m}{An odd \code{\link{integer}} giving the number of adjacent points to be used.}

\item{p}{An \code{\link{integer}} giving the degree of the polynomial to be used.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Savitzky-Golay Filter
}
\details{
This method is based on the least-squares fitting of polynomials to
segments of \eqn{m} adjacent points.

There will be \eqn{(m - 1) / 2} points both at the beginning and at the end
of the spectrum for which a complete \eqn{m}-width smooth cannot be
calculated. To prevent data loss, the original \eqn{(m - 1) / 2} points at
the ends of the spectrum are preserved.
}
\examples{
## gamma-ray spectrometry
data("LaBr")

## Plot spectrum
LaBr <- signal_select(LaBr, from = 1350, to = 1650)
plot(LaBr, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Rectangular smoothing
LaBr_unweighted <- smooth_rectangular(LaBr, m = 3)
plot(LaBr_unweighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Triangular smoothing
LaBr_weighted <- smooth_triangular(LaBr, m = 5)
plot(LaBr_weighted, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Loess smoothing
LaBr_loess <- smooth_loess(LaBr, span = 0.75)
plot(LaBr_loess, type = "l", xlab = "Energy (keV)", ylab = "Count")

## Savitzky–Golay filter
LaBr_savitzky <- smooth_savitzky(LaBr, m = 21, p = 2)
plot(LaBr_savitzky, type = "l", xlab = "Energy (keV)", ylab = "Count")
}
\references{
Gorry, P. A. (1990). General Least-Squares Smoothing and Differentiation by
the Convolution (Savitzky-Golay) Method. \emph{Analytical Chemistry}, 62(6),
p. 570-573. \doi{10.1021/ac00205a007}.

Savitzky, A. & Golay, M. J. E. (1964). Smoothing and Differentiation of
Data by Simplified Least Squares Procedures. \emph{Analytical Chemistry},
36(8), p. 1627-1639. \doi{10.1021/ac60214a047}.
}
\seealso{
Other smoothing methods: 
\code{\link{smooth_loess}()},
\code{\link{smooth_rectangular}()},
\code{\link{smooth_triangular}()}
}
\author{
N. Frerebeau
}
\concept{smoothing methods}
