% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/signal.R
\docType{methods}
\name{signal_transform}
\alias{signal_transform}
\alias{signal_transform-method}
\alias{signal_transform,numeric,numeric-method}
\alias{signal_transform,ANY,missing-method}
\title{Transform Intensities}
\usage{
signal_transform(x, y, ...)

\S4method{signal_transform}{numeric,numeric}(x, y, f, ...)

\S4method{signal_transform}{ANY,missing}(x, f, ...)
}
\arguments{
\item{x, y}{A \code{\link{numeric}} vector. If \code{y} is missing, an attempt is made to
interpret \code{x} in a suitable way (see \code{\link[grDevices:xy.coords]{grDevices::xy.coords()}}).}

\item{...}{Extra arguments to be passed to \code{f}.}

\item{f}{A \code{\link{function}} that takes a \code{numeric} vector of intensities as
argument and returns a \code{numeric} vector.}
}
\value{
Returns a \code{\link{list}} with two components \code{x} and \code{y}.
}
\description{
Transform Intensities
}
\details{
The stabilization step aims at improving the identification of peaks with a
low signal-to-noise ratio. This particularly targets higher energy peaks.
}
\seealso{
Other signal processing methods: 
\code{\link{smooth}()},
\code{\link{subset}()}
}
\author{
N. Frerebeau
}
\concept{signal processing methods}
