% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alkahest-package.R
\docType{package}
\name{alkahest-package}
\alias{alkahest-package}
\alias{_PACKAGE}
\alias{alkahest}
\title{alkahest: Pre-Processing XY Data from Experimental Methods}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A lightweight, dependency-free toolbox for pre-processing XY data from experimental methods (i.e. any signal that can be measured along a continuous variable). This package provides methods for baseline estimation and correction, smoothing, normalization, integration and peaks detection. Baseline correction methods includes polynomial fitting as described in Lieber and Mahadevan-Jansen (2003) \doi{10.1366/000370203322554518}, Rolling Ball algorithm after Kneen and Annegarn (1996) \doi{10.1016/0168-583X(95)00908-6}, SNIP algorithm after Ryan et al. (1988) \doi{10.1016/0168-583X(88)90063-8}, 4S Peak Filling after Liland (2015) \doi{10.1016/j.mex.2015.02.009} and more.
}
\details{
\tabular{ll}{
\strong{Package:} \tab alkahest \cr
\strong{Type:} \tab Package \cr
\strong{Version:} \tab 1.1.0 \cr
\strong{License:} \tab GPL-3 \cr
\strong{Zenodo:} \tab \doi{10.5281/zenodo.7081524} \cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://packages.tesselle.org/alkahest/}
  \item \url{https://github.com/tesselle/alkahest}
  \item Report bugs at \url{https://github.com/tesselle/alkahest/issues}
}

}
\author{
\strong{Full list of authors and contributors} (alphabetic order)

\tabular{ll}{
Nicolas Frerebeau \tab \emph{Université Bordeaux Montaigne, France} \cr
Brice Lebrun \tab \emph{Université Bordeaux Montaigne, France} \cr
}

\strong{Package maintainer}

Nicolas Frerebeau\cr
\email{nicolas.frerebeau@u-bordeaux-montaigne.fr}

Archéosciences Bordeaux (UMR 6034)\cr
Maison de l'Archéologie\cr
Université Bordeaux Montaigne\cr
33607 Pessac cedex\cr
France
}
\keyword{internal}
