% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaBr}
\alias{LaBr}
\title{Gamma-Ray Spectrometry}
\format{
A \code{\link{data.frame}} with 1024 rows (channels) and 2 variables.
\describe{
\item{energy}{(keV)}
\item{count}{}
}
}
\usage{
LaBr
}
\description{
Gamma-Ray Spectrometry
}
\examples{
data("LaBr")
plot(LaBr, type = "l", xlab = "Energy (keV)", ylab = "Count")
}
\seealso{
Other datasets: 
\code{\link{BEGe}},
\code{\link{XRD}}
}
\concept{datasets}
\keyword{datasets}
